requirements.txt
requests==2.32.3
ollama
pandas
fastapi
langchain
langchain-community
yacana
filetype
rapidocr-onnxruntime
pypandoc
pymupdf
officeparserpy
openpyxl
xlrd
llama-index-embeddings-ollama
llama-index-llms-ollama
llama-index-vector-stores-faiss
pathlib
pydantic
faiss-cpu
tqdm
tomli